Update ltspice.py
I have tried to load a raw file generated with ngspice-31, and it failed with a raised FileSizeNotMatchException. Reason was, that ngspice (at least in the version I have used) does not write 'double' in the flags section, but 'real', resulting in a mismatch in the data section calculation by a factor of 4. This proposal for code modification may help to provide better backward compatibility to older ngspice versions.